+++ system/src/Grav/Common/User/FlexUser/Storage/UserFileStorage.php
@@ -13,4 +13,13 @@ use Grav\Framework\Flex\Storage\FileStorage;
 
 class UserFileStorage extends FileStorage
 {
+    /**
+     * {@inheritdoc}
+     * @see FlexStorageInterface::getMediaPath()
+     */
+    public function getMediaPath(string $key = null): ?string
+    {
+        // There is no media support for file storage (fallback to common location).
+        return null;
+    }
 }