+++ system/src/Grav/Framework/Flex/Interfaces/FlexStorageInterface.php
@@ -133,16 +133,16 @@ interface FlexStorageInterface
      *
      * @param  string|null $key Optional storage key.
      *
-     * @return string Path in the filesystem. Can be URI.
+     * @return string|null Path in the filesystem. Can be URI or null if storage is not filesystem based.
      */
-    public function getStoragePath(string $key = null): string;
+    public function getStoragePath(string $key = null): ?string;
 
     /**
      * Get filesystem path for the collection or object media.
      *
      * @param  string|null $key Optional storage key.
      *
-     * @return string Path in the filesystem. Can be URI.
+     * @return string|null Path in the filesystem. Can be URI or null if media isn't supported.
      */
-    public function getMediaPath(string $key = null): string;
+    public function getMediaPath(string $key = null): ?string;
 }